refactor: extract chart painter and utility functions for better code organization
- Move AnimatedChartPainter from widget file to separate painter file
- Extract common utility functions to lib/src/core/util/helper.dart
- Extract color utilities to lib/src/core/util/colors.dart
- Remove duplicate code and improve maintainability
- Reduce animated_chart_widget.dart from ~2,200 to ~900 lines
- All tests passing, no functional changes
